SF: make channel field zeroy double (not float) to ensure it is saved correctly in config validators

+++ src/ui/scoring-facility.hh
@@ -218,7 +218,7 @@ class SScoringFacility {
 				return _h;
 			}
 
-		float	zeroy;
+		double	zeroy;
 		bool operator<( const SChannel& rv) const;
 
 		double	signal_display_scale;
